🚀 chore(README.md): update package name and installation instructions The package name has been updated from 'userefreshtoken' to 'use-refresh-token' to better reflect its purpose. The installation instructions have been updated accordingly.
🔨 chore(README.md): add handleRefreshToken function to documentation
The handleRefreshToken function has been added to the documentation to explain its purpose and how it can be used.

🔨 chore(index.d.ts): add handleRefreshToken function to TFunctions type
The handleRefreshToken function has been added to the TFunctions type to allow it to be used in the options object.

🔨 chore(index.ts): add handleRefreshToken to options object
The handleRefreshToken function has been added to the options object to allow it to be passed in during installation.